BigQuery Pet-project
As a data analyst I extracted and analyzed data to evaluate key metrics such as CR (conversion rate) and correlation. You can investigate the calculation of conversions in terms of dates and traffic channels, comparison of conversion between different landing pages and testing the correlation between user engagement and purchases. The project can be useful for marketers to choose a marketing strategy, sales channels and optimize advertising costs.